Move index configuration files to separate package
