+++ textrepo-app/src/main/java/nl/knaw/huc/config/TextRepoConfiguration.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.dropwizard.Configuration;
 import io.dropwizard.db.DataSourceFactory;
 import io.federecio.dropwizard.swagger.SwaggerBundleConfiguration;
-import nl.knaw.huc.service.index.MappedIndexerConfiguration;
+import nl.knaw.huc.service.index.config.MappedIndexerConfiguration;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
