+++ textrepo-app/src/main/java/nl/knaw/huc/service/index/TextRepoElasticClient.java
@@ -1,5 +1,6 @@
 package nl.knaw.huc.service.index;
 
+import nl.knaw.huc.service.index.config.ElasticsearchConfiguration;
 import org.apache.http.HttpHost;
 import org.elasticsearch.client.RestHighLevelClient;
