+++ textrepo-app/src/main/java/nl/knaw/huc/service/index/config/MappedIndexerConfiguration.java
@@ -1,6 +1,7 @@
-package nl.knaw.huc.service.index;
+package nl.knaw.huc.service.index.config;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import nl.knaw.huc.service.index.FieldsConfiguration;
 
 public class MappedIndexerConfiguration extends IndexerConfiguration {
