+++ textrepo-app/src/test/java/nl/knaw/huc/service/index/MappedIndexerTest.java
@@ -3,11 +3,8 @@ package nl.knaw.huc.service.index;
 import nl.knaw.huc.core.TextRepoFile;
 import nl.knaw.huc.core.Type;
 import nl.knaw.huc.service.type.TypeService;
-import nl.knaw.huc.service.index.MappedIndexerConfiguration;
-import nl.knaw.huc.service.index.IndexerException;
-import nl.knaw.huc.service.index.MappedIndexer;
-import nl.knaw.huc.service.index.ElasticsearchConfiguration;
-import nl.knaw.huc.service.index.FieldsConfiguration;
+import nl.knaw.huc.service.index.config.MappedIndexerConfiguration;
+import nl.knaw.huc.service.index.config.ElasticsearchConfiguration;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
